+++ pages/affiliation/visualizations/commit_domains.py
@@ -8,22 +8,22 @@ import logging
 from dateutil.relativedelta import *  # type: ignore
 import plotly.express as px
 from pages.utils.graph_utils import color_seq
-from queries.company_query import company_query as cmq
+from queries.commits_query import commits_query as cq
 import io
 from cache_manager.cache_manager import CacheManager as cm
 from pages.utils.job_utils import nodata_graph
 import time
 import datetime as dt
 
-PAGE = "company"
-VIZ_ID = "company-core-contributors"
+PAGE = "affiliation"
+VIZ_ID = "commit-domains"
 
-gc_company_core_contributors = dbc.Card(
+gc_commit_domains = dbc.Card(
     [
         dbc.CardBody(
             [
                 html.H3(
-                    "Company Core Contributors",
+                    "Commit Activity by Domain",
                     className="card-title",
                     style={"textAlign": "center"},
                 ),
@@ -31,12 +31,13 @@ gc_company_core_contributors = dbc.Card(
                     [
                         dbc.PopoverHeader("Graph Info:"),
                         dbc.PopoverBody(
-                            "This graph counts the number of core contributions that COULD be linked to each company.\n\
-                            The methodology behind this is to take each associated email to someones github account\n\
-                            and link the contributions to each as it is unknown which initity the actvity was done for.\n\
-                            Then the graph groups contributions by contributors and filters by contributors that are core.\n\
-                            Contributions required is the amount of contributions necessary to be consider a core contributor\n\
-                            Core Contributors required is the amount of core contributors needed to have the domain listed."
+                            """
+                            Visualizes the proportion of commit activity done by specific email domains.\n
+                            e.g. if there are 100 commits and 75 commits were authored by a contributor with a\n
+                            '@gmail.com' email address, 75 percent of the chart will be represented '@gmail.com.'\n
+                            This can help to capture the relative magnitude of commit contribution by various corporate\n
+                            or institutional entities.
+                            """
                         ),
                     ],
                     id=f"popover-{PAGE}-{VIZ_ID}",
@@ -53,37 +54,21 @@ gc_company_core_contributors = dbc.Card(
                             [
                                 dbc.Label(
                                     "Contributions Required:",
-                                    html_for=f"contributions-required-{PAGE}-{VIZ_ID}",
+                                    html_for=f"company-contributions-required-{PAGE}-{VIZ_ID}",
                                     width={"size": "auto"},
                                 ),
                                 dbc.Col(
-                                    dbc.Input(
-                                        id=f"contributions-required-{PAGE}-{VIZ_ID}",
-                                        type="number",
-                                        min=1,
-                                        max=100,
-                                        step=1,
-                                        value=10,
-                                        size="sm",
-                                    ),
-                                    className="me-2",
-                                    width=2,
-                                ),
-                                dbc.Label(
-                                    "Core Contributors Required:",
-                                    html_for=f"contributors-required-{PAGE}-{VIZ_ID}",
-                                    width={"size": "auto"},
-                                ),
-                                dbc.Col(
-                                    dbc.Input(
-                                        id=f"contributors-required-{PAGE}-{VIZ_ID}",
-                                        type="number",
-                                        min=1,
-                                        max=50,
-                                        step=1,
-                                        value=3,
-                                        size="sm",
-                                    ),
+                                    [
+                                        dbc.Input(
+                                            id=f"company-contributions-required-{PAGE}-{VIZ_ID}",
+                                            type="number",
+                                            min=1,
+                                            max=100,
+                                            step=1,
+                                            value=10,
+                                            size="sm",
+                                        ),
+                                    ],
                                     className="me-2",
                                     width=2,
                                 ),
@@ -141,20 +126,19 @@ def toggle_popover(n, is_open):
     Output(f"{PAGE}-{VIZ_ID}", "figure"),
     [
         Input("repo-choices", "data"),
-        Input(f"contributions-required-{PAGE}-{VIZ_ID}", "value"),
-        Input(f"contributors-required-{PAGE}-{VIZ_ID}", "value"),
+        Input(f"company-contributions-required-{PAGE}-{VIZ_ID}", "value"),
         Input(f"date-picker-range-{PAGE}-{VIZ_ID}", "start_date"),
         Input(f"date-picker-range-{PAGE}-{VIZ_ID}", "end_date"),
     ],
     background=True,
 )
-def compay_associated_activity_graph(repolist, contributions, contributors, start_date, end_date):
+def commit_domains_graph(repolist, num, start_date, end_date):
     # wait for data to asynchronously download and become available.
     cache = cm()
-    df = cache.grabm(func=cmq, repos=repolist)
+    df = cache.grabm(func=cq, repos=repolist)
     while df is None:
         time.sleep(1.0)
-        df = cache.grabm(func=cmq, repos=repolist)
+        df = cache.grabm(func=cq, repos=repolist)
 
     start = time.perf_counter()
     logging.warning(f"{VIZ_ID}- START")
@@ -165,7 +149,7 @@ def compay_associated_activity_graph(repolist, contributions, contributors, star
         return nodata_graph
 
     # function for all data pre processing, COULD HAVE ADDITIONAL INPUTS AND OUTPUTS
-    df = process_data(df, contributions, contributors, start_date, end_date)
+    df = process_data(df, num, start_date, end_date)
 
     fig = create_figure(df)
 
@@ -173,28 +157,23 @@ def compay_associated_activity_graph(repolist, contributions, contributors, star
     return fig
 
 
-def process_data(df: pd.DataFrame, contributions, contributors, start_date, end_date):
-    # convert to datetime objects rather than strings
-    df["created"] = pd.to_datetime(df["created"], utc=True)
+def process_data(df: pd.DataFrame, num, start_date, end_date):
+    # TODO: create docstring
 
-    # order values chronologically by COLUMN_TO_SORT_BY date
-    df = df.sort_values(by="created", axis=0, ascending=True)
+    # convert to datetime objects rather than strings
+    df["author_timestamp"] = pd.to_datetime(df["author_timestamp"], utc=True)
+
+    # order values chronologically by author_timestamp date earliest to latest
+    df = df.sort_values(by="author_timestamp", axis=0, ascending=True)
 
     # filter values based on date picker
     if start_date is not None:
-        df = df[df.created >= start_date]
+        df = df[df.author_timestamp >= start_date]
     if end_date is not None:
-        df = df[df.created <= end_date]
+        df = df[df.author_timestamp <= end_date]
 
-    # groups contributions by countributor id and counts, created column now hold the number
-    # of contributions for its respective contributor
-    df = df.groupby(["cntrb_id", "email_list"], as_index=False)[["created"]].count()
-
-    # filters out contributors that dont meet the core contribution threshhold
-    df = df[df.created >= contributions]
-
-    # creates list of unique emails and flattens list result
-    emails = df.email_list.str.split(" , ").explode("email_list").tolist()
+    # creates list of emails for each contribution and flattens list result
+    emails = df.author_email.tolist()
 
     # remove any entries not in email format
     emails = [x for x in emails if "@" in x]
@@ -205,17 +184,17 @@ def process_data(df: pd.DataFrame, contributions, contributors, start_date, end_
     # creates df of domains and counts
     df = pd.DataFrame(email_domains, columns=["domains"]).value_counts().to_frame().reset_index()
 
-    df = df.rename(columns={0: "contributors"})
+    df = df.rename(columns={0: "occurrences"})
 
     # changes the name of the company if under a certain threshold
-    df.loc[df.contributors <= contributors, "domains"] = "Other"
+    df.loc[df.occurrences <= num, "domains"] = "Other"
 
     # groups others together for final counts
     df = (
-        df.groupby(by="domains")["contributors"]
+        df.groupby(by="domains")["occurrences"]
         .sum()
         .reset_index()
-        .sort_values(by=["contributors"], ascending=False)
+        .sort_values(by=["occurrences"], ascending=False)
         .reset_index(drop=True)
     )
 
@@ -224,17 +203,11 @@ def process_data(df: pd.DataFrame, contributions, contributors, start_date, end_
 
 def create_figure(df: pd.DataFrame):
     # graph generation
-    fig = px.bar(df, x="domains", y="contributors", color_discrete_sequence=color_seq)
-    fig.update_xaxes(rangeslider_visible=True, range=[-0.5, 15])
-    fig.update_layout(
-        xaxis_title="Domains",
-        yaxis_title="Core Contributors",
-        bargroupgap=0.1,
-        margin_b=40,
-        font=dict(size=14),
-    )
+    fig = px.pie(df, names="domains", values="occurrences", color_discrete_sequence=color_seq)
     fig.update_traces(
-        hovertemplate="%{label} <br>Contributors: %{value}<br><extra></extra>",
+        textposition="inside",
+        textinfo="percent+label",
+        hovertemplate="%{label} <br>Commits: %{value}<br><extra></extra>",
     )
 
     return fig
