+++ pages/affiliation/visualizations/unqiue_domains.py
@@ -15,15 +15,15 @@ from pages.utils.job_utils import nodata_graph
 import time
 import datetime as dt
 
-PAGE = "company"
-VIZ_ID = "company-associated-activity"
+PAGE = "affiliation"
+VIZ_ID = "unique-domains"
 
-gc_company_associated_activity = dbc.Card(
+gc_unique_domains = dbc.Card(
     [
         dbc.CardBody(
             [
                 html.H3(
-                    "Company Associated Activity",
+                    "Unique Contributor Email Domains",
                     className="card-title",
                     style={"textAlign": "center"},
                 ),
@@ -32,19 +32,10 @@ gc_company_associated_activity = dbc.Card(
                         dbc.PopoverHeader("Graph Info:"),
                         dbc.PopoverBody(
                             """
-                            For non-commit contributions (see definition of Contribution on Info page)\n
-                            we only know which contributor account contributed. We don't know which email,\n
-                            and therefore which possible institution the contribution represented.\n
-                            e.g. if we know that a PR comment was made by JaneDoe, and they have a '@redhat.com' and\n
-                            an '@gmail.com' email, we don't know whether they contributed individually\n
-                            or as a representative of an instituion. Therefore, we lower-bound the contribution\n
-                            of representation by counting each contribution as being made by ALL of the contributors\n
-                            linked email domains.\n
-                            This graph can therefore be interpreted as 'The minimum number of individuals who have\n
-                            been associated with each domain.'\n
-                            e.g. If there are 100 contributions and 20 contributors, and each contributor has an '@redhat.com'\n
-                            email associated with their account and one other random email, '@redhat.com' will be counted 100 times\n
-                            and the other contributor emails will also total a count of 100.\n
+                            Visualizes the population of unique commit email addresses per represented domain.\n
+                            e.g. if there are 100 distinct commit contributors and 50 use an '@gmail.com' email address,\n
+                            and another 50 use an '@redhat.com' email address, 50 percent of of emails wll be '@gmail.com'\n
+                            and 50% will be '@redhat.com'.
                             """
                         ),
                     ],
@@ -61,18 +52,18 @@ gc_company_associated_activity = dbc.Card(
                         dbc.Row(
                             [
                                 dbc.Label(
-                                    "Contributions Required:",
-                                    html_for=f"company-contributions-required-{PAGE}-{VIZ_ID}",
+                                    "Contributors Required:",
+                                    html_for=f"contributions-required-{PAGE}-{VIZ_ID}",
                                     width={"size": "auto"},
                                 ),
                                 dbc.Col(
                                     dbc.Input(
-                                        id=f"company-contributions-required-{PAGE}-{VIZ_ID}",
+                                        id=f"contributions-required-{PAGE}-{VIZ_ID}",
                                         type="number",
                                         min=1,
-                                        max=100,
+                                        max=50,
                                         step=1,
-                                        value=10,
+                                        value=3,
                                         size="sm",
                                     ),
                                     className="me-2",
@@ -132,26 +123,13 @@ def toggle_popover(n, is_open):
     Output(f"{PAGE}-{VIZ_ID}", "figure"),
     [
         Input("repo-choices", "data"),
-        Input(f"company-contributions-required-{PAGE}-{VIZ_ID}", "value"),
+        Input(f"contributions-required-{PAGE}-{VIZ_ID}", "value"),
         Input(f"date-picker-range-{PAGE}-{VIZ_ID}", "start_date"),
         Input(f"date-picker-range-{PAGE}-{VIZ_ID}", "end_date"),
     ],
     background=True,
 )
-def compay_associated_activity_graph(repolist, num, start_date, end_date):
-    """Each contribution is associated with a contributor. That contributor can be associated with
-
-    more than one different email. Hence each contribution is associated with all of the emails that a contributor has historically used.
-
-    We don't always know which email (and therefore which organization) a contributor is affiliated with at contribution
-
-    time, so we choose to count all of their possible affiliations via their email list. e.g. if "Jane Doe" is associated with "gmail.com"
-
-    and "yahoo.com" and they have 5 contributions, "gmail.com" and "yahoo.com" would be counted 5 times each. We assume that relatively few people
-
-    will have many emails. We acknowledge that this will almost always contribute to an overcount but will never undercount."
-    """
-
+def unique_domains_graph(repolist, num, start_date, end_date):
     # wait for data to asynchronously download and become available.
     cache = cm()
     df = cache.grabm(func=cmq, repos=repolist)
@@ -189,8 +167,8 @@ def process_data(df: pd.DataFrame, num, start_date, end_date):
     if end_date is not None:
         df = df[df.created <= end_date]
 
-    # creates list of emails for each contribution and flattens list result
-    emails = df.email_list.str.split(" , ").explode("email_list").tolist()
+    # creates list of unique emails and flattens list result
+    emails = df.email_list.str.split(" , ").explode("email_list").unique().tolist()
 
     # remove any entries not in email format
     emails = [x for x in emails if "@" in x]
@@ -201,17 +179,17 @@ def process_data(df: pd.DataFrame, num, start_date, end_date):
     # creates df of domains and counts
     df = pd.DataFrame(email_domains, columns=["domains"]).value_counts().to_frame().reset_index()
 
-    df = df.rename(columns={0: "occurrences"})
+    df = df.rename(columns={0: "occurences"})
 
     # changes the name of the company if under a certain threshold
-    df.loc[df.occurrences <= num, "domains"] = "Other"
+    df.loc[df.occurences <= num, "domains"] = "Other"
 
     # groups others together for final counts
     df = (
-        df.groupby(by="domains")["occurrences"]
+        df.groupby(by="domains")["occurences"]
         .sum()
         .reset_index()
-        .sort_values(by=["occurrences"], ascending=False)
+        .sort_values(by=["occurences"], ascending=False)
         .reset_index(drop=True)
     )
 
@@ -220,16 +198,10 @@ def process_data(df: pd.DataFrame, num, start_date, end_date):
 
 def create_figure(df: pd.DataFrame):
     # graph generation
-    fig = px.bar(df, x="domains", y="occurrences", color_discrete_sequence=color_seq)
-    fig.update_xaxes(rangeslider_visible=True, range=[-0.5, 15])
-    fig.update_layout(
-        xaxis_title="Domains",
-        yaxis_title="Contributions",
-        bargroupgap=0.1,
-        margin_b=40,
-        font=dict(size=14),
-    )
+    fig = px.pie(df, names="domains", values="occurences", color_discrete_sequence=color_seq)
     fig.update_traces(
+        textposition="inside",
+        textinfo="percent+label",
         hovertemplate="%{label} <br>Contributions: %{value}<br><extra></extra>",
     )
