+++ pages/contributions/visualizations/commits_over_time.py
@@ -3,25 +3,25 @@ import dash
 import dash_bootstrap_components as dbc
 from dash import callback
 from dash.dependencies import Input, Output, State
-import plotly.graph_objects as go
 import pandas as pd
 import logging
+import plotly.express as px
 from pages.utils.graph_utils import get_graph_time_values, color_seq
-from pages.utils.job_utils import nodata_graph
-from queries.issues_query import issues_query as iq
+from queries.commits_query import commits_query as cmq
 from cache_manager.cache_manager import CacheManager as cm
+from pages.utils.job_utils import nodata_graph
 import io
 import time
 
-PAGE = "overview"
-VIZ_ID = "issues-over-time"
+PAGE = "contributions"
+VIZ_ID = "commits-over-time"
 
-gc_issues_over_time = dbc.Card(
+gc_commits_over_time = dbc.Card(
     [
         dbc.CardBody(
             [
                 html.H3(
-                    "Issues Over Time",
+                    "Commits Over Time",
                     className="card-title",
                     style={"textAlign": "center"},
                 ),
@@ -30,13 +30,13 @@ gc_issues_over_time = dbc.Card(
                         dbc.PopoverHeader("Graph Info:"),
                         dbc.PopoverBody(
                             """
-                            Visualizes the activity of issues being Opened and Closed paritioned by time-window.\n
-                            Also shows the total volume of Open issues over time.
+                            Visualizes the number of commits added to the project.\n
+                            Commits are counted relative to a user-selected time window.
                             """
                         ),
                     ],
                     id=f"popover-{PAGE}-{VIZ_ID}",
-                    target=f"popover-target-{PAGE}-{VIZ_ID}",
+                    target=f"popover-target-{PAGE}-{VIZ_ID}",  # needs to be the same as dbc.Button id
                     placement="top",
                     is_open=False,
                 ),
@@ -105,7 +105,7 @@ def toggle_popover(n, is_open):
     return is_open
 
 
-# callback for issues over time graph
+# callback for commits over time graph
 @callback(
     Output(f"{PAGE}-{VIZ_ID}", "figure"),
     [
@@ -114,40 +114,37 @@ def toggle_popover(n, is_open):
     ],
     background=True,
 )
-def issues_over_time_graph(repolist, interval):
+def commits_over_time_graph(repolist, interval):
     # wait for data to asynchronously download and become available.
     cache = cm()
-    df = cache.grabm(func=iq, repos=repolist)
+    df = cache.grabm(func=cmq, repos=repolist)
     while df is None:
         time.sleep(1.0)
-        df = cache.grabm(func=iq, repos=repolist)
+        df = cache.grabm(func=cmq, repos=repolist)
 
     # data ready.
     start = time.perf_counter()
-    logging.warning("ISSUES OVER TIME - START")
+    logging.warning("COMMITS_OVER_TIME_VIZ - START")
 
     # test if there is data
     if df.empty:
-        logging.warning("ISSUES OVER TIME - NO DATA AVAILABLE")
+        logging.warning("COMMITS OVER TIME - NO DATA AVAILABLE")
         return nodata_graph
 
     # function for all data pre processing
-    df_created, df_closed, df_open = process_data(df, interval)
+    df_created = process_data(df, interval)
 
-    fig = create_figure(df_created, df_closed, df_open, interval)
+    fig = create_figure(df_created, interval)
 
-    logging.warning(f"ISSUES_OVER_TIME_VIZ - END - {time.perf_counter() - start}")
-
+    logging.warning(f"COMMITS_OVER_TIME_VIZ - END - {time.perf_counter() - start}")
     return fig
 
 
 def process_data(df: pd.DataFrame, interval):
-    # convert to datetime objects rather than strings
-    df["created"] = pd.to_datetime(df["created"], utc=True)
-    df["closed"] = pd.to_datetime(df["closed"], utc=True)
-
-    # order values chronologically by creation date
-    df = df.sort_values(by="created", axis=0, ascending=True)
+    # convert to datetime objects with consistent column name
+    # incoming value should be a posix integer.
+    df["date"] = pd.to_datetime(df["date"], utc=True)
+    df.rename(columns={"date": "created"}, inplace=True)
 
     # variable to slice on to handle weekly period edge case
     period_slice = None
@@ -155,74 +152,35 @@ def process_data(df: pd.DataFrame, interval):
         # this is to slice the extra period information that comes with the weekly case
         period_slice = 10
 
-    # data frames for issues created or closed. Detailed description applies for all 3.
-
-    # get the count of created issues in the desired interval in pandas period format, sort index to order entries
-    created_range = pd.to_datetime(df["created"]).dt.to_period(interval).value_counts().sort_index()
-
-    # converts to data frame object and creates date column from period values
-    df_created = created_range.to_frame().reset_index().rename(columns={"index": "Date"})
+    # get the count of commits in the desired interval in pandas period format, sort index to order entries
+    df_created = (
+        df.groupby(by=df.created.dt.to_period(interval))["commits"]
+        .nunique()
+        .reset_index()
+        .rename(columns={"created": "Date"})
+    )
 
     # converts date column to a datetime object, converts to string first to handle period information
     # the period slice is to handle weekly corner case
     df_created["Date"] = pd.to_datetime(df_created["Date"].astype(str).str[:period_slice])
 
-    # df for closed issues in time interval
-    closed_range = pd.to_datetime(df["closed"]).dt.to_period(interval).value_counts().sort_index()
-    df_closed = closed_range.to_frame().reset_index().rename(columns={"index": "Date"})
-    df_closed["Date"] = pd.to_datetime(df_closed["Date"].astype(str).str[:period_slice])
-
-    # formatting for graph generation
-    if interval == "M":
-        df_created["Date"] = df_created["Date"].dt.strftime("%Y-%m-01")
-        df_closed["Date"] = df_closed["Date"].dt.strftime("%Y-%m-01")
-    elif interval == "Y":
-        df_created["Date"] = df_created["Date"].dt.strftime("%Y-01-01")
-        df_closed["Date"] = df_closed["Date"].dt.strftime("%Y-01-01")
-
-    # first and last elements of the dataframe are the
-    # earliest and latest events respectively
-    earliest = df["created"].min()
-    latest = max(df["created"].max(), df["closed"].max())
-
-    # beginning to the end of time by the specified interval
-    dates = pd.date_range(start=earliest, end=latest, freq="D", inclusive="both")
-
-    # df for open issues for time interval
-    df_open = dates.to_frame(index=False, name="Date")
-
-    # aplies function to get the amount of open issues for each day
-    df_open["Open"] = df_open.apply(lambda row: get_open(df, row.Date), axis=1)
-
-    df_open["Date"] = df_open["Date"].dt.strftime("%Y-%m-%d")
-
-    return df_created, df_closed, df_open
+    return df_created
 
 
-def create_figure(df_created: pd.DataFrame, df_closed: pd.DataFrame, df_open: pd.DataFrame, interval):
+def create_figure(df_created: pd.DataFrame, interval):
     # time values for graph
     x_r, x_name, hover, period = get_graph_time_values(interval)
 
-    # graph generation
-    fig = go.Figure()
-    fig.add_bar(
-        x=df_created["Date"],
-        y=df_created["created"],
-        opacity=0.9,
-        hovertemplate=hover + "<br>Created: %{y}<br>" + "<extra></extra>",
-        offsetgroup=0,
-        marker=dict(color=color_seq[0]),
-        name="Created",
+    # graph geration
+    fig = px.bar(
+        df_created,
+        x="Date",
+        y="commits",
+        range_x=x_r,
+        labels={"x": x_name, "y": "Commits"},
+        color_discrete_sequence=[color_seq[3]],
     )
-    fig.add_bar(
-        x=df_closed["Date"],
-        y=df_closed["closed"],
-        opacity=0.9,
-        hovertemplate=hover + "<br>Closed: %{y}<br>" + "<extra></extra>",
-        offsetgroup=1,
-        marker=dict(color=color_seq[3]),
-        name="Closed",
-    )
+    fig.update_traces(hovertemplate=hover + "<br>Commits: %{y}<br>")
     fig.update_xaxes(
         showgrid=True,
         ticklabelmode="period",
@@ -232,36 +190,10 @@ def create_figure(df_created: pd.DataFrame, df_closed: pd.DataFrame, df_open: pd
     )
     fig.update_layout(
         xaxis_title=x_name,
-        yaxis_title="Number of Issues",
-        bargroupgap=0.1,
+        yaxis_title="Number of Commits",
         margin_b=40,
+        margin_r=20,
         font=dict(size=14),
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=df_open["Date"],
-            y=df_open["Open"],
-            mode="lines",
-            marker=dict(color=color_seq[5]),
-            name="Open",
-            hovertemplate="Issues Open: %{y}<br>%{x|%b %d, %Y} <extra></extra>",
-        )
     )
 
     return fig
-
-
-# for each day, this function calculates the amount of open issues
-def get_open(df, date):
-    # drop rows that are more recent than the date limit
-    df_lim = df[df["created"] <= date]
-
-    # drops rows that have been closed after date
-    df_open = df_lim[df_lim["closed"] > date]
-
-    # include issues that have not been close yet
-    df_open = pd.concat([df_open, df_lim[df_lim.closed.isnull()]])
-
-    # generates number of columns ie open issues
-    num_open = df_open.shape[0]
-    return num_open
