+++ pages/contributions/visualizations/issue_staleness.py
@@ -5,25 +5,27 @@ from dash import callback
 from dash.dependencies import Input, Output, State
 import plotly.graph_objects as go
 import pandas as pd
+import datetime as dt
 import logging
 from dateutil.relativedelta import *  # type: ignore
 import plotly.express as px
 from pages.utils.graph_utils import get_graph_time_values, color_seq
+from queries.issues_query import issues_query as iq
 from pages.utils.job_utils import nodata_graph
-from queries.prs_query import prs_query as prq
-import time
-import io
 from cache_manager.cache_manager import CacheManager as cm
+import io
+import time
 
-PAGE = "overview"
-VIZ_ID = "pr-staleness"
+PAGE = "contributions"
+VIZ_ID = "issue-staleness"
 
-gc_pr_staleness = dbc.Card(
+
+gc_issue_staleness = dbc.Card(
     [
         dbc.CardBody(
             [
                 html.H3(
-                    "Pull Request Activity- Staleness",
+                    "Issue Activity- Staleness",
                     className="card-title",
                     style={"textAlign": "center"},
                 ),
@@ -32,14 +34,14 @@ gc_pr_staleness = dbc.Card(
                         dbc.PopoverHeader("Graph Info:"),
                         dbc.PopoverBody(
                             """
-                            Visualizes growth of Open Pull Request backlog. Differentiates sub-populations\n
-                            of PRs by their 'Staleness.'\n
+                            Visualizes growth of Issue backlog. Differentiates sub-populations\n
+                            of issues by their 'Staleness.'\n
                             Please see the definition of 'Staleness' on the Info page.
                             """
                         ),
                     ],
                     id=f"popover-{PAGE}-{VIZ_ID}",
-                    target=f"popover-target-{PAGE}-{VIZ_ID}",  # needs to be the same as dbc.Button id
+                    target=f"popover-target-{PAGE}-{VIZ_ID}",
                     placement="top",
                     is_open=False,
                 ),
@@ -86,16 +88,16 @@ gc_pr_staleness = dbc.Card(
                                     className="me-2",
                                     width=1,
                                 ),
+                                dbc.Alert(
+                                    children="Please ensure that 'Days Until Staling' is less than 'Days Until Stale'",
+                                    id=f"check-alert-{PAGE}-{VIZ_ID}",
+                                    dismissable=True,
+                                    fade=False,
+                                    is_open=False,
+                                    color="warning",
+                                ),
                             ],
                             align="center",
-                        ),
-                        dbc.Alert(
-                            children="Please ensure that 'Days Until Staling' is less than 'Days Until Stale'",
-                            id=f"check-alert-{PAGE}-{VIZ_ID}",
-                            dismissable=True,
-                            fade=False,
-                            is_open=False,
-                            color="warning",
                         ),
                         dbc.Row(
                             [
@@ -162,7 +164,7 @@ def toggle_popover(n, is_open):
     ],
     background=True,
 )
-def new_staling_prs_graph(repolist, interval, staling_interval, stale_interval):
+def new_staling_issues_graph(repolist, interval, staling_interval, stale_interval):
     # conditional for the intervals to be valid options
     if staling_interval > stale_interval:
         return dash.no_update, True
@@ -172,17 +174,17 @@ def new_staling_prs_graph(repolist, interval, staling_interval, stale_interval):
 
     # wait for data to asynchronously download and become available.
     cache = cm()
-    df = cache.grabm(func=prq, repos=repolist)
+    df = cache.grabm(func=iq, repos=repolist)
     while df is None:
         time.sleep(1.0)
-        df = cache.grabm(func=prq, repos=repolist)
+        df = cache.grabm(func=iq, repos=repolist)
 
     start = time.perf_counter()
-    logging.warning("PULL REQUEST STALENESS - START")
+    logging.warning("ISSUES STALENESS - START")
 
     # test if there is data
     if df.empty:
-        logging.warning("PULL REQUEST STALENESS  - NO DATA AVAILABLE")
+        logging.warning("ISSUE STALENESS - NO DATA AVAILABLE")
         return nodata_graph, False
 
     # function for all data pre processing
@@ -190,14 +192,13 @@ def new_staling_prs_graph(repolist, interval, staling_interval, stale_interval):
 
     fig = create_figure(df_status, interval)
 
-    logging.warning(f"PULL REQUEST STALENESS - END - {time.perf_counter() - start}")
+    logging.warning(f"ISSUE STALENESS - END - {time.perf_counter() - start}")
     return fig, False
 
 
 def process_data(df: pd.DataFrame, interval, staling_interval, stale_interval):
     # convert to datetime objects rather than strings
     df["created"] = pd.to_datetime(df["created"], utc=True)
-    df["merged"] = pd.to_datetime(df["merged"], utc=True)
     df["closed"] = pd.to_datetime(df["closed"], utc=True)
 
     # order values chronologically by creation date
@@ -211,7 +212,7 @@ def process_data(df: pd.DataFrame, interval, staling_interval, stale_interval):
     # generating buckets beginning to the end of time by the specified interval
     dates = pd.date_range(start=earliest, end=latest, freq=interval, inclusive="both")
 
-    # df for new, staling, and stale prs for time interval
+    # df for new, staling, and stale issues for time interval
     df_status = dates.to_frame(index=False, name="Date")
 
     # dynamically apply the function to all dates defined in the date_range to create df_status
@@ -245,7 +246,7 @@ def create_figure(df_status: pd.DataFrame, interval):
                     y=df_status["New"],
                     mode="lines",
                     showlegend=True,
-                    hovertemplate="PRs New: %{y}<br>%{x|%b %d, %Y} <extra></extra>",
+                    hovertemplate="Issues New: %{y}<br>%{x|%b %d, %Y} <extra></extra>",
                     marker=dict(color=color_seq[0]),
                 ),
                 go.Scatter(
@@ -254,7 +255,7 @@ def create_figure(df_status: pd.DataFrame, interval):
                     y=df_status["Staling"],
                     mode="lines",
                     showlegend=True,
-                    hovertemplate="PRs Staling: %{y}<br>%{x|%b %d, %Y} <extra></extra>",
+                    hovertemplate="Issues Staling: %{y}<br>%{x|%b %d, %Y} <extra></extra>",
                     marker=dict(color=color_seq[3]),
                 ),
                 go.Scatter(
@@ -263,7 +264,7 @@ def create_figure(df_status: pd.DataFrame, interval):
                     y=df_status["Stale"],
                     mode="lines",
                     showlegend=True,
-                    hovertemplate="PRs Stale: %{y}<br>%{x|%b %d, %Y} <extra></extra>",
+                    hovertemplate="Issues Stale: %{y}<br>%{x|%b %d, %Y} <extra></extra>",
                     marker=dict(color=color_seq[5]),
                 ),
             ]
@@ -277,11 +278,11 @@ def create_figure(df_status: pd.DataFrame, interval):
         )
 
         # edit hover values
-        fig.update_traces(hovertemplate=hover + "<br>PRs: %{y}<br>" + "<extra></extra>")
+        fig.update_traces(hovertemplate=hover + "<br>Issues: %{y}<br>" + "<extra></extra>")
 
     fig.update_layout(
         xaxis_title="Time",
-        yaxis_title="Pull Requests",
+        yaxis_title="Issues",
         legend_title="Type",
         font=dict(size=14),
     )
@@ -305,10 +306,10 @@ def get_new_staling_stale_up_to(df, date, staling_interval, stale_interval):
     # time difference for the amount of days before the threshold date
     stale_days = date - relativedelta(days=+stale_interval)
 
-    # PRs still open at the specified date
+    # issuess still open at the specified date
     numTotal = df_in_range.shape[0]
 
-    # num of currently open PRs that have been create in the last staling_value amount of days
+    # num of currently open issues that have been create in the last staling_value amount of days
     numNew = df_in_range[df_in_range["created"] >= staling_days].shape[0]
 
     staling = df_in_range[df_in_range["created"] > stale_days]
