+++ pages/contributors/visualizations/contrib_activity_cycle.py
